Renamed the outgoing physcial mutator to a better name
It was already changed for v6 so not a breaking change
