+++ src/NServiceBus.AcceptanceTests/MessageId/When_message_has_no_id_header.cs
@@ -4,6 +4,7 @@
     using NServiceBus.AcceptanceTesting;
     using NServiceBus.AcceptanceTests.EndpointTemplates;
     using NServiceBus.Config;
+    using NServiceBus.MessageMutator;
     using NServiceBus.Pipeline;
     using NServiceBus.Unicast.Messages;
     using NUnit.Framework;
@@ -26,7 +27,7 @@
             Assert.IsFalse(string.IsNullOrWhiteSpace(context.MessageId));
         }
 
-        public class CorruptionMutator : IMutateOutgoingPhysicalContext
+        public class CorruptionMutator : IMutateOutgoingPhysicalMessages
         {
             public Context ScenarioContext { get; set; }
 
@@ -35,7 +36,7 @@
              
             }
 
-            public void MutateOutgoing(OutgoingPhysicalMutatorContext context)
+            public void MutateOutgoing(MutateOutgoingPhysicalMessageContext context)
             {
                 context.SetHeader(Headers.MessageId,null);
                 context.SetHeader("ScenarioContextId",ScenarioContext.Id.ToString());
